requirements.txt
crewai==0.1.0
crewai-tools==0.1.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-openai==0.0.5
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.0.0
PyYAML==6.0.1
jinja2==3.1.2
email-validator==2.1.0
tqdm==4.66.1
sqlalchemy==2.0.0
ratelimit==2.2.1
beautifulsoup4==4.12.2
requests==2.31.0